requirements.txt
torch>=2.0.0
transformers>=4.30.0
pytorch-lightning>=2.0.0
datasets>=2.12.0
pyyaml>=6.0
click>=8.0.0
seqeval>=1.2.2
pytorch-crf>=0.7.2
python-dotenv
